requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
torch>=2.0.0
